Practice Files/Function_Integrations/Function_Integrations/requirements.txt
azure-functions
azure.storage.blob